+++ RELEASE-NOTES.md
@@ -11,6 +11,13 @@ VERSION SCHEMA: vYYYY.MM.#SEQ
 
 ---
 
+## v2019.05.2
+
+- NEW: Add Register-Criminal-Fingerprints-For-Present-Person workflow.
+- NEW: Add Register-Criminal-Fingerprints-For-Not-Present-Person workflow.
+
+---
+
 ## v2019.05.1
 
 - FIX: Using civil biometrics ID and criminal biometrics ID in register-convicted-report-for-non-present workflow is not complete.